raises salary test passes

+++ src/main/java/Management/Manager.java
@@ -6,7 +6,7 @@ public class Manager extends Employee {
 
     String deptName;
 
-    public Manager(String name, String niNumber, double salary, String deptName){
+    public Manager(String name, String niNumber, int salary, String deptName){
         super(name, niNumber, salary);
         this.deptName = deptName;
     }
